+++ webdev/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Add an explicit error if there are no directories to serve. Typically this
   would happen if the user doesn't have a `web` directory.
 - Add support for specifying `--hostname any`.
+- DevTools no longer launches in a new window. This prevents an issue where
+  pausing your application also pauses DevTools.
 
 ## 2.0.7